SimpleSAML_Session: Update isValid() to require a authority string.
This patch updates the isValid function to require a valid authority.
All users of isValid in the simpleSAMLphp source pass the authority
parameter. If anyone uses isValid() without a valid authority they
must update their source code.

This change is made to make it more difficult to use simpleSAMLphp in
an insecure way.