InfoCard: Fix insecure file creation.
The InfoCard modules stores files in /tmp, but does this in a way that
allows another user with local access to the machine to overwrite any
file the webserver has access to.

This patch fixes it by moving the files into the simpleSAMLphp temp
directory.

Thanks to Thijs Kinkhorst <thijs@uvt.nl> for reporting this bug.

Merged into 1.5 branch from r2082.